refactor(pilot): extract polling setup to poller_setup.go (GH-1214)
Move runPollingMode (~1,064 lines) and its helpers (applyInputOverrides,
applyTeamOverrides) into poller_setup.go. This removes ~1,100 lines from
main.go, making the file more manageable.

These functions are called from newStartCmd and handle the entire
polling-only mode orchestration including adapter setup, autopilot
controllers, and brief scheduling.
